Adding comments to the project

+++ src/component/Login/index.js
@@ -1,27 +1,32 @@
 import './index.css'
 import React, { useState } from 'react';
-import { getFirestore, doc, getDoc } from 'firebase/firestore'
+import { doc, getDoc } from 'firebase/firestore'
 
-export default function Login({app}){
+export default function Login({db}){
 
-    const db = getFirestore(app);
     const [email, setEmail] = useState('');
     const [password, setPassword] = useState('');
     
+      // Handles the login functionality of the user
     async function handleLogin (e){
         e.preventDefault()
         
-        const ref = doc(db, "users", email); //"Bring me, from the collection 'users' the document with name/value 'email'"
+        // We create a doc that 'points' at collection 'users' with primary key user's input email 
+        const ref = doc(db, "users", email); 
+        // Now "Bring me, from the collection 'users' the document with name/value 'email'"
         const res = await getDoc(ref);
 
-        //If the user with email = "email" exists in the db...
-        if (res.exists()) {
+        //If the user with email = "email" and password = "passowrd" exists in the db...
+        if (res.exists() && res.data().email === email && res.data().password === password) {
+            // Get the role and email...
             const user_role = res.data().role
             const user_email = res.data().email
 
+            // Store the email and role as keys in your browser local storage
             localStorage.setItem('role', user_role)
             localStorage.setItem('email', user_email)
 
+            // Go to page /courses
             window.location.href = './courses'
             console.log("Found User:", res.data());
         } else {
